充值功能

+++ app/src/main/java/com/tiyi/tiyi_app/dto/CoinModel.kt
@@ -28,7 +28,7 @@ data class RechargeItemsModel(
         @SerialName("id") val id: Int,
         @SerialName("title") val title: String,
         @SerialName("amount") val amount: Int,
-        @SerialName("price") val price: Int
+        @SerialName("price") val price: Double
     )
 }
